+++ datasource/src/main/scala/astraea/spark/rasterframes/datasource/geotrellis/SpatialFilterPushdownRules.scala
@@ -20,11 +20,12 @@
 package astraea.spark.rasterframes.datasource.geotrellis
 
 import astraea.spark.rasterframes._
+import astraea.spark.rasterframes.datasource.geotrellis.GeoTrellisRelation.FilterPredicate
 import astraea.spark.rasterframes.expressions.SpatialExpression.{Intersects, RelationPredicate}
 import com.vividsolutions.jts.geom.Geometry
 import geotrellis.util.LazyLogging
 import org.apache.spark.sql.SQLRules.GeometryLiteral
-import org.apache.spark.sql.catalyst.expressions.AttributeReference
+import org.apache.spark.sql.catalyst.expressions.{AttributeReference, Literal}
 import org.apache.spark.sql.catalyst.plans.logical.{Filter, LogicalPlan}
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.execution.datasources.LogicalRelation
@@ -44,23 +45,32 @@ import org.apache.spark.sql.execution.datasources.LogicalRelation
 object SpatialFilterPushdownRules extends Rule[LogicalPlan] with LazyLogging {
   import astraea.spark.rasterframes.encoders.GeoTrellisEncoders._
 
-  case class FilterPredicate(colName: String, relation: RelationPredicate, geom: Geometry)
+  val EXTENT_COL_NAME = EXTENT_COLUMN.columnName
 
   object ExtentAttr {
     def unapply(in: AttributeReference): Boolean =
-      in.name == EXTENT_COLUMN.columnName &&
+      in.name == EXTENT_COL_NAME &&
       in.dataType == extentEncoder.schema
   }
 
   def apply(plan: LogicalPlan): LogicalPlan = {
     plan.transform {
-      case f @ Filter(condition, lr @ LogicalRelation(gm: GeoTrellisRelation, _, _)) ⇒
-        condition match {
+      case f @ Filter(condition, lr @ LogicalRelation(gt: GeoTrellisRelation, _, _)) ⇒
+        val newFilter = condition match  {
           case i @ Intersects(ExtentAttr(), g: GeometryLiteral) ⇒
-            // Introduce SFCurve relation
-            f
-          case _ ⇒ f
+            Some(FilterPredicate(EXTENT_COL_NAME, i.nodeName, g.geom))
+          case i @ Intersects(ExtentAttr(), Literal(geom: Geometry, _)) ⇒
+            Some(FilterPredicate(EXTENT_COL_NAME, i.nodeName, geom))
+          case _ ⇒ None
         }
+
+        // Only add the filter if it doesn't already exist.
+        // Lenses would clean this up a bit.
+        newFilter.filterNot(gt.filters.contains)
+          .map(gt.withFilter)
+          .map(r ⇒ lr.copy(relation = r))
+          .map(lr ⇒ Filter(condition, lr))
+          .getOrElse(f)
     }
   }
 }