+++ src/main/scala/astraea/spark/rasterframes/expressions/SpatialExpression.scala
@@ -19,18 +19,15 @@
 
 package astraea.spark.rasterframes.expressions
 
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
-import org.apache.spark.sql.types._
 import astraea.spark.rasterframes.encoders.GeoTrellisEncoders._
 import astraea.spark.rasterframes.expressions.SpatialExpression.RelationPredicate
 import astraea.spark.rasterframes.jts.SpatialEncoders._
 import com.vividsolutions.jts.geom._
-import geotrellis.vector.GeometryCollection
-import org.apache.spark.sql.{GeometryUDT, PointUDT, SQLSpatialFunctions}
-import org.geotools.geometry.jts.JTSFactoryFinder
-import org.locationtech.geomesa.curve.XZ2SFC
+import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.catalyst.expressions._
+import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
+import org.apache.spark.sql.types._
+import org.apache.spark.sql._
 
 
 /**
@@ -59,10 +56,19 @@ abstract class SpatialExpression extends BinaryExpression with CodegenFallback {
             jtsPointEncoder.resolveAndBind().fromRow(r)
           case t if t.getClass.isAssignableFrom(PointUDT.getClass) ⇒
             PointUDT.deserialize(r)
+          case t if t.getClass.isAssignableFrom(MultiPointUDT.getClass) ⇒
+            MultiPointUDT.deserialize(r)
+          case t if t.getClass.isAssignableFrom(LineStringUDT.getClass) ⇒
+            LineStringUDT.deserialize(r)
+          case t if t.getClass.isAssignableFrom(MultiLineStringUDT.getClass) ⇒
+            MultiLineStringUDT.deserialize(r)
+          case t if t.getClass.isAssignableFrom(PolygonUDT.getClass) ⇒
+            PolygonUDT.deserialize(r)
+          case t if t.getClass.isAssignableFrom(MultiPolygonUDT.getClass) ⇒
+            MultiPolygonUDT.deserialize(r)
           case t if t.getClass.isAssignableFrom(GeometryUDT.getClass) ⇒
             GeometryUDT.deserialize(r)
         }
-      case sfcIndex: Long ⇒ SpatialExpression.Index(sfcIndex)
     }
   }
 
@@ -78,23 +84,6 @@ abstract class SpatialExpression extends BinaryExpression with CodegenFallback {
 
 object SpatialExpression {
   type RelationPredicate = (Geometry, Geometry) ⇒ java.lang.Boolean
-  // TODO: externalize
-  val xzPrecision = 20.toShort
-
-  val sfc = XZ2SFC(xzPrecision)
-
-  @transient
-  private val geomFactory = JTSFactoryFinder.getGeometryFactory
-
-  private case class Index(sfcIndex: Long) extends Point(
-    geomFactory.getCoordinateSequenceFactory.create(0, 0), geomFactory
-  ) {
-    override def intersects(g: Geometry): Boolean = {
-      val env = g.getEnvelopeInternal
-      val ranges = sfc.ranges(env.getMinX, env.getMinY, env.getMaxX, env.getMaxY)
-      ranges.exists(r ⇒ r.lower <= sfcIndex && sfcIndex <= r.upper)
-    }
-  }
 
   case class Intersects(left: Expression, right: Expression) extends SpatialExpression {
     override def nodeName = "intersects"